requirements.txt
openai
python-dotenv
markitdown[pdf, docx, pptx]